requirements.txt
django<1.11
bs4
gunicorn
ipdb
markdown
pilkit
pillow
psycopg2
pytz
six
translitcodec
unicodecsv
# unicodedata2 is a backport for python2
unicodedata2


django-admin-sortable2
django-admin-steroids
django-content-editor
django-extensions
# django-imagekit
# django-imperavi
django-markdown
#django-markup
django-model-utils
django-mptt
django-polymorphic
django-reversion
# solid_i18n
django-solo
django-sortedm2m
django-stronghold

-e git+gogs@projects.c--y.net:erik/django-shared-utils.git#egg=utils
-e git+gogs@projects.c--y.net:erik/django-shared-multilingual.git#egg=django_shared_multilingual
